Introduce `dist` target
This adds the `dist` target to the Makefile that's meant to build
binaries for all supported OS/ARCH combinations and let's the `build`
target build only for the local OS/ARCH. Which makes for a quicker
turnaround when one only needs to build locally.

Also the Makefile has been refactored slightly to remove pattern-based
target, so now it is a bit more manageable.

The list of supported OS/ARCH is fetched from golang: OS/ARCH
combinations that have first class support in golang minus i386
architecture will count as supported.