Acceptance test coverage report
This modifies the source files to add code coverage instructions much
the same as `go tool cover` would do using
github.com/mendersoftware/gobinarycoverage. The `acceptance` target now
copies the source code to a temporary directory makes the source code
instrumentation and applies the patch to make sure the `coverReport` is
invoked and that the missing `os` import is added. Next the build is
performed so that the instrumented version of the `ec` command is built
for use by the acceptance tests. Which in turn configure environment
variables to generate the `coverage-acceptance-*.out` file in the root
of the repository.