Accept that wiremock might not be used
When the WireMock client is not used, either by the WireMock not being
able to start or it has not used at all in a scenario, we should not
fail when trying to fetch it from the Context as it won't be there.

In case of not using WireMock at all this is expected and we should not
proceed in the After hook.

In case of error starting WireMock the panic when not being able to find
the client could shadow the error presented at starting the WireMock in
the first place.