Assert the standard output in the acceptance tests
This adds the ability to assert standard output in acceptance tests.
This way we can catch changes to the standard output to maintain
backward compatibility in a fairly visible way.

The output is presented in diff format and if the terminal supports it
in color, for example:

```
--- FAIL: TestFeatures/happy_day (6.84s)
    suite.go:449: expected and actual output differ:
         {
           "components": [
             0: {
        -      "containerImage": "localhost:WAT/acceptance/ec-happy-day",
        +      "containerImage": "localhost:49339/acceptance/ec-happy-day",
               "name": "Unnamed",
               "success": true,
               "violations": [
                 0: {
        -          "filename": "/tmp/ecp_input.(\d+)/input.json",
        +          "filename": "/tmp/ecp_input.2534943934/input.json",
                   "namespace": "release.main",
                   "successes": 0
                 }
               ]
             }
           ],
           "success": true
         }
```

Given that the JSON diff compares values exactly, values that do match
by the regular expression will still be printed. But matching all values
either by exact values or by regular expressions will not fail the test
and will not display the diff.