Fix deadlock in eval command
Prior to this commit, if there were any errors found during the
goroutine to validate an image, nothing was written to the component
channel. Since this channel is always read immediatelly after calling
the goroutine, it blocked forever waiting for something to be added to
the component channel.

This commit fixes the issue by ensuring something is always written to
the component channel.

https://issues.redhat.com/browse/HACBS-759

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>