Refactor eval command
This refactors the eval command to process both a single image and an
ApplicationSnapshot parameters in the same manner. When image is
specified, an artificial ApplicationSnapshot is created containing one
Unnamed component with the provided image.
The bulk of the processing is pulled out into two functions to ease
maintenance and testing.