Restore stubbed environment in acceptance tests
When tests are run with `-persist` parameter a `.persisted` file is
created with data needed to re-run the acceptance tests with the exactly
the same data as it was run. To re-run the acceptance tests in such a
way `-restore` parameter needs to be specified.

With this a marker interface is introduced `testenv.WithState` needed
only to access the key under which the state is stored in
`context.Context` and in the persisted file.

Given that the same structure is used, for ease of development, for
storing data in `context.Context` and in the persisted file, storing
complex values (such as `v1.Image` from `google/go-containerregistry`)
is not possible, the data needs to be fetched. For example in the case
of signature/attestation data it is fetched from the stubbed registry
and not stored in memory within `context.Context`.

Using a single key to store the whole persisted data in
`context.Context` reduces the computational complexity from O(N) to
O(1); though there are still some values that are stored in
`context.Context` and not managed by `testenv`, such as the ones in the
`cli` package.

The `registry.go` was moved to the `registry` package in order for the
`key` type to differ between it and `image` package.