Fix Accept header parsing
In acceptance tests when a scenario finishes to help stub API requests
towards stubbed services (currently k8s and rekor), we dump all
unstubbed (unmatched by WireMock) requests. For the output we also print
the content type of the request which we parse using the `mime` package.
Since `Accept` header can be multi-valued
(e.g. `Accept: application/json, */*`) we need to parse and split on `,`