Vendor acceptance test coverage tool
The tool we use to instrument the source code in order to gather the
acceptance test coverage writes to standard error from the `coverReport`
method that is inserted into the `main.go`. This interferes with any
assertion of the `ec` command line stderr output.
This pulls in the https://github.com/mendersoftware/gobinarycoverage
tool into `internal/acceptance/coverage/coverage.go` and removes those
stderr outputs from the template used to generate the code
instrumentation.
In addition to that, the changes we used to keep in
`patches/acceptance_cover.diff` are now added to the coverage tool, so
we don't need to carry/maintain or, of course, apply that patch to run
the acceptance tests.