Refine JSON regex matching
When comparing two JSON trees in acceptance tests a case was missing to
handle `*gojsondiff.Modified` delta type, this adds handling of that
case and adds a test to verify the behavior. Also when matching with
regular expressions we need to make sure that we're matching the correct
positions, so the clause matching `TextDiff` before `PostData` was
removed as it could match with unpaired positions. This also makes sure
that array cardinality issues are caught.