Add debug messages for improved troubleshooting
I originally planned to add warnings and errors as well as debug
messages, but decide to hold off until there's a clear plan for how
we want error handling to work.

It might makes sense to use log.Error or log.Fatal for catching and
reporting errors but let's figure out the details in a future patch.

These debug messages have proven useful already to me already.
Before this patch it was pretty hard to figure out where and how
things were going wrong. Now you can turn on debug and get a pretty
good idea from reading the debug messages.

This isn't meant to be an exhaustive list of debug output, but
should be a good start. (It's probably better to merge early and
improve iteratively rather than debate the choice of debug messages
and their wording.)