Realistic Rekor stub in acceptance tests
This adds more data to the responses from the stubbed Rekor service, now
containing the tree and verification information; and expands to include
image signature attestation in addition to the build attestation.

Since we're using an ad-hoc Rekor signing key we need to provide it to
cosign embedded within the CLI using the `SIGSTORE_REKOR_PUBLIC_KEY`
environment variable. This also most likely points to an improvement we
need in the `ec`, i.e. we need to provide the Rekor public key to it.

The stub response is now tied to the request by the means of log entry
data (build attestation or image signature), and no longer returns the
single entry for all requests to the stub.