Hide Conftest implementation details
This introduces the `Evaluator` interface so we don't depend (so much)
on the implementation of `conftestEvaluator`, which is now private with
private fields. As `context.Context` can vary between calls, e.g. if we
want to run image and attestation signature validation concurrently it
needs to be passed as function parameter rather than stored in the
evaluator implementation.