Refactor acceptance tests for robustness
Introduces `IsRunning` function to several stubs so it is easier to
handle scenarios where those stubs are not running. Fetching data about
non-running stubs leads to errors and this introduces a flag we can use
to prevent that.
For robustness we can also check if the Context has a value in it or not
via `HasState` function, to make sure we don't `panic` if the state
hasn't been setup, e.g. when steps that would setup state are not
invoked (used). Common case is in the After functions which might try to
inspect the state even though the state wasn't setup, such as in the
wiremock stub.
Also includes minor refactor for fetching the testing.T instance from
the Context. And refactoring the `ec` invocation step in `cli.go` by
splitting it over several functions for readability.