Skip caching in lint workflow
The setup-go and lint workflows both try to cache the same files, this
leads to issues when the golangci-lint action tries to restore the cache
as it fails because files already exist as they have been restored by
setup-go. Since setup-go caches the package and build files we don't
need the golangci-lint action to cache at all.