Merge acceptance tests coverage reports
Acceptance tests run `ec` that has been instrumented to generate the
coverage report multiple times, which produces multiple code coverage
reports. The `mv` in the Makefile would attempt to move several files
over to one file which would fail with errors like:

```
 mv: target '/home/runner/work/ec-cli/ec-cli/coverage-acceptance.out' is not a directory
```

This uses a tool to merge the coverage reports into a single instead of
the crude `mv` command that only worked when we had one invocation of
the `ec` in acceptance tests.