Deterministic output from `track bundle` command
This uses `yamlfmt` to format the output of the `track bundle` command,
which sorts the keys resulting in a deterministic output.

Property based test that asserts that the output is sorted has been
added. Given the nature of generative tests the test timeout was
increased from 500ms to 30s to accommodate many samples being generated
and test cases being run. It would complicate this change to introduce
separation of short and long tests, so in lieu of that it was simpler to
increase the timeout, this is not something we should tolerate for long
and we should revert to 500ms test deadline with exceptions for long
tests. There is a proposal to add per-test timeout deadlines, but that
might be far in the future for us.