Explicitly download all go dependencies
I've noticed that the lint workflow still downloads some go
dependencies. Ideally we would like to get all dependencies from the
cache and have the cache populated with all needed dependencies. What we
currently have is that the cache is populated with dependencies on
demand, and since the cache key is generated from the `go.sum` file
those partial dependencies will be cached and the cache will not
updated unless `go.sum` changes, and even then with partial dependencies
from the running workflow.
This explicitly downloads all dependencies in all workflows so we have
the cache populated with everything that we need.