Make sure we use `ec-cli` for content
Accessing the first component of the Antora site works okay when there
is only one component, when there are more than one or the content can
be added over time we should not rely on choosing the content by it's
index. This explicitly tries to find the `ec-cli` component instead.