Setup publishing to npmjs.com instead of GitHub
Now we shall publish only to npmjs.com and not to GitHub NPM Packages
Registry. This makes it more convenient as the clients of the package
don't need to authenticate against the NPM registry that hosts the
package.
Also took the opportunity to set the Antora dependency as a development
dependency.