Fix ec-replace diff generation
Use os.exec instead of go-git to generate the diff. This is not ideal,
but there are two issues with using go-git for this use case.

First, it is currently not possible to produce a diff for uncommitted
changes. Discussion: https://github.com/src-d/go-git/issues/604

Second, the output generated by go-git is just plain wrong and in some
cases does not work with "git apply".

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>