Add support for time-based policies
Time-based policies are defined as policies that are only effective at a
given point in time. Prior to that time, a failure to meet the policy is
reported as a warning instead.

This logic currently resides in the ec-policies repository. This commit
adds the logic to the ec-cli in an effort to simplify the logic in the
policies repo.

Prior to removing the logic from ec-policies, it's important that all
users are upgraded to a version of the ec-cli containing this commit.
It's ok if both places are performing this logic which results in a
no-op in the ec-cli. But at least one of them must do so.

https://issues.redhat.com/browse/HACBS-1101

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>