HACBS-1089 Report errors if no predicate.materials
This commit adds functionality that tracks whether attestations provide
a `predicate.materials` entry. If no attestations for a component
provide a valid `predicate.materials` value, we return an error for that
component which is added to a multierror object. After all components
have been evaluated, we return our multierror object, indicating if we
have any errors found.

Signed-off-by: Rob Nester <rnester@redhat.com>