Support fetching data from a git source
As mentioned in the comments, it's meant to be a temporary solution
until we have better support for defining sources that are data
rather than rego files.

This implementation looks for the prescence of "/data" in the url.
If it's found it assumes the source is meant to be data instead of
rego files, and it downloads it to the data directory instead of the
policy directory so that conftest handles it correctly.

See also the next commit where this is used to download the existing
acceptable bundles and rule collection data.

JIRA: HACBS-1206