Add one special hard-coded data source for data
Originally, when we were checking out the entire ec-policies repo it
included the top level data directory where the allowed bundle list
was defined, and more recently the rule collections list. Now that
we're fetching just the policy directory, the data needs to be fetch
separately.

Rather than requiring each ECP configuration to add the data source,
hard code it so it is always fetched. This idea of hard coding some
data sources in the ec-cli is maybe not a good long term solution,
and might be replaced with something else in future.

JIRA: HACBS-1206