Error conventions linter & fixer
Adds a linter/fixer for the convention of not returning raw errors.
Raw errors are errors that haven't been wrapped by our `pkg/error`, i.e.
they don't have an error code and our custom message.

Currently the linter doesn't fail in GitHub but reports issues as
annotations. And the suggestions are not applied when running
`make lint-fix`.