Use k8s.io/client-go directly, configure logs
To reduce the depenencies a little we can use k8s.io/client-go directly
and not via sigs.k8s.io/controller-runtime. The klog logs are now
forwarded to logrus instead of going to stdout. The caveat being that
all klog logs are logged with INFO category, though they do comply with
the logging level set, as the filtering is done to make sure the correct
logging level is propagated -- the sink bridge logs with `Info`
regardless.