Evaluate each policy source independently
This changes the policy evaluation to evaluate each policy source
independently, instead of downloading from all policy sources and then
evaluate once with using all policy files from all policy sources.

In addition to that handling of the hard coded data source from the
https://github.com/hacbs-contract/ec-policies is now in the evaluator
package so that the downloader package has a single responsibility -- to
download from the provided sources.

Along with that the logic of where to place the policy files has been
moved to the source package for the same reason.

Ref. https://issues.redhat.com/browse/HACBS-1311