Fix the TUF cache concurrency issue
It could be that the reason why the Checks workflow failed in some of
the recent runs:

 - https://github.com/hacbs-contract/ec-cli/actions/runs/3439351522
 - https://github.com/hacbs-contract/ec-cli/actions/runs/3443912318
 - https://github.com/hacbs-contract/ec-cli/actions/runs/3443914333
 - https://github.com/hacbs-contract/ec-cli/actions/runs/3444243298

As we run the acceptance test features concurrently the `ec` command
line is invoked concurrently. That makes it prone to a race condition
when accessing the filesystem based TUF cache that the Sigstore
framework is trying to create/update.

This disables the filesystem based cache and uses the in-memory cache
instead.