Introduce the internal/tools submodule
The versions of tools and especially their dependencies might collide
with the versions we want to use in ec-cli. Also it is best to keep a
minimum of dependencies in ec-cli. With that, this introduces a
submodule in `internal/tools` to keep track of versions of the tools we
use. Now instead of `go run ...` we need to use
`go run -modfile internal/tools/go.mod ...`