`hack/setup-environment.sh` improvements
Now the `hack/setup-environment.sh` in addition to configuring Tekton
Pipelines and the EnterpriseContractPolicy custom resource and
permissions, also configures an image registry that the ec-cli image and
the verify-enterprise-contract Tekton Task bundle can be pushed to.

The versions of the Tekton Pipeline and EnterpriseContractPolicy custom
resource are governed by the versions used in the `go.mod` file.

Additionally a namespace named `work` is created where the required
permissions are granted to the `default` ServiceAccount.

For each component (Tekton Pipeline, ECC, registry, work) a kustomize
directory is created to keep the descriptors together and help with
further customization.