Use unique subdir for hard coded data source
This prevents the data source from overwriting the config.json that
was created a moment earlier in the data directory. To do this
nicely, a new way to explictly distinguish between policy sources
and data sources is introduced.

"Policy" policy sources contain rego files and are downloaded to a
unique subdir under $workdir/policy. "Data" policy sources contain
json and yaml files and are downloaded to a unique subdir under
$workdir/data.

This is hopefully a step towards what is needed for HACBS-1396.

Ref: https://issues.redhat.com/browse/HACBS-1421