Build ec-cli images for multiple architectures
This builds the ec-cli image for arm64 and ppc64le in addition to amd64.

Additional changes to the image are that ubi-minimal is used, reducing
the image by 1/3 (~100MB) and the ENTRYPOINT is specified.

To build on foreign architectures qemu-static must be installed, e.g. on
Fedora `qemu-user-static-aarch64` and `qemu-user-static-ppc`.