Workaround for k8s client initialization in tkn
For `tkn bundle push` a k8s client is initialized, if the command is run
without a valid Kubernetes configuration it'll fail.

This provides a fake Kubernetes configuration file, and using the
`KUBECONFIG` environment variable configures the k8s client to use it.

See https://github.com/tektoncd/cli/pull/1807