Development environment with Tekton Chains
Adds `hack/setup-dev-environment.sh` that will create a Kind cluster
named `ec` if it already doesn't exist, install Tekton Pipielines,
Tekton Chains, ECC, an image registry and setup a "work" namespace with
the required permissions.

Also adds a `dev` target to the Makefile to push the ec and the task
image to the development cluster.