Simplify the kustomize test setup
Replaces the use of `kubectl kustomize` with invoking the `kustomize`
directly (version managed by `internal/tools/go.mod`). Instead of
scripting in the `hack/setup-environment.sh` to determine the versions
of Tekton and enterprise-contract-controller, that is now performed by
running KRM functions via kustomize generators. This opens a way to
install all required resources using a single kustomize command and a
kustomization that references individual components.