Fix unit test when kubeconfig is NOT invalid
The unit test is verifying an error is returning when creating the
client in an environment where a kubeconfig is not available.

However, this test now fails if KUBECONFIG is set in the enironment
running the tests.

Instead of messing with the HOME environment variable, the commit makes
the test leverage the KUBECONFIG environment variable to point to a path
that does not exist. KUBECONFIG has a higher precendence.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>