Automated PR update of the Tekton Task bundle
Adds automated creation of pull requests in infra-deployments when a new
Tekton Task bundle is released. The prerequisite is that the
`EC Automation`[1] GitHub App is installed in the redhat-appstudio
GitHub organization and enabled for the infra-deployments repository.
And that the deploy key is set in the `DEPLOY_KEY` secret for the fork
repository `hacbs-contract/infra-deployments`.

The installation ID (`APP_INSTALL_ID`) needs to be maintained with any
changes.

[1] https://github.com/organizations/hacbs-contract/settings/apps/ec-automation