Split checks and release cache
It could be that the long build times for `make dist` run in the release
workflow is due to the cache misses for architectures/oses other than
amd64/linux which is run in the checks workflow which is the first
workflow to run and maintain the cache key against the `go.sum` files,
not taking into the account that we build for more architectures/oses in
the release workflow.

This splits the checks (amd64/linux) and the release caches (*/*) so
that they're separately maintained.