Nicer version command output
Adds more data to the version command output. Including, a pseudo
version in the form
`v0.1.<number of commits on this branch>-<commit id>`, the change
date of the last commit, and version of some components used.

Version can be outputted in `--short` or `--json` format.

Ref. https://issues.redhat.com/browse/HACBS-1502