Organize cmd package
The cmd package is now organized into sub-package (and sub-directories)
that match the CLI structure.

This also moves the create-pr and replace commands under the
experimental sub-command to better represent their state.

The `fs` and `withFs` helper functions have been moved to the generic
internal/utils package to avoid cyclic imports. It also seems to be a
better home for it as it may be used in other parts of the code base.

https://issues.redhat.com/browse/HACBS-1736

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>