Don't set "origin" for generated antora files
Antora is expecting origin to be an object, or nothing. Using a
string causes this inscrutable message to appear instead of the real
error:

  Cannot use 'in' operator to search for 'worktree' in reference

Take a look at getFileLocation in
node_modules/@antora/content-classifier/lib/content-catalog.js if
you want to know why.

IIRC the origin value is related to the "edit this file source"
links that antora likes to create. If we do need to supress those
for generated content, I'm sure there are other ways to do it.