Convert result "collections" Metadata to []string
The change in #465 converted the `result.Metadata["collections"]` to
`[]string` from `[]any` leading to `extractCollections` returning no
collections for a given rule. This meant that when filtering by
collections was used no failures would be reported.

This changes `result.Metadata["collections"]` to always be `[]string`.
Any other given value type will cause `collections` to be removed from
`result.Metadata".

Side note: this was not detected by unit tests because the tests did not
provide any `.rego` files for the rule inspection to inspect, so no
conversion from `[]any` to `[]string` was performed in `addMetadata`.
And we have no acceptance tests that specifically test rule filtering by
collections.