Include successful rule outcomes in output
In addition to failures and warnings now successes are added to the
output. Given that from Conftest we only receive the number of
successful rule outcomes this relies on the inspected AST data from the
policy sources.

We can't tell the message a successful rule would generate, this is lost
when Conftest discards this and doesn't include it in the result, a
fixed message of `"Pass"` is reported.

Ref. https://issues.redhat.com/browse/HACBS-1791

Co-authored-by: Simon Baird <simon.baird@gmail.com>