Add new INFO param to the ec task definition
The param sets the --info flag and is enabled by default.

The motivation is so that the titles and descriptions for all rules,
including passing checks as well as violations and warnings, are
exposed in the log output of the task where they can (presumably) be
scraped by the UI team and used to produce user-facing displays of
the EC results.

Previously the titles and descriptions were not available in the ec
output, so they needed to be slurped in from a static json file. Now
the information is available directly in the ec output, the json
file should not be needed any more.