Task should not fail on warnings in strict mode
The new jq test from #572 fails the task when running with the
strict param set, unless the result is "SUCCESS", which means that a
result of "WARNING" is now failing unexpectedly. Noticed by
@shebert.

The possible values are:
- SUCCESS
- WARNING
- FAILURE
- SKIPPED
- ERROR (part of the HACBS_TEST_OUTPUT spec, but not used by ec currently)

Here we're saying that everything other than SUCCESS and WARNING
will cause the task to fail.

I'm not entirely sure whether SKIPPED should count as a failure or a
pass when we're in strict mode. We could change our mind later
maybe.