Extract solution annotation & include in report
It shows up if you use the --info flag.

As per the commentary I decided not to include it in the success
output. This is demonstrated by the fact that the "Easy" solution in
the acceptance test example rego doesn't show up in the output for
the always passing test when ec is run with the --info flag.

JIRA: HACBS-2049