Configure Ingress via Nginx
This does take up ports 80 and 443 on all network interfaces. But allows
accessing fulcio via `fulcio.localhost`, e.g.

    curl -k https://fulcio.localhost

The wildcard `*.localhost` certificate for the ingress is issued by
tricking Kubernetes to sign it as if it was a node to be added to the
cluster. This makes it a bit easier, as we don't need to add
cert-manager or the like and can trust one CA root certificate for the
whole ordeal. For example:

    SSL_CERT_FILE=<(kubectl get cm kube-root-ca.crt -o jsonpath="{['data']['ca\.crt']}") cosign sign-blob