Add acceptance tests for keyless workflow
https://issues.redhat.com/browse/HACBS-2071

An image signature/attesttaion created by cosign contains a
SignedEntityTimestamp (SET). This includes all the information needed
for verfication if the TUF root has already been initialized.

To initialize a TUF root, two parameters are needed. The TUF mirror and
the trusted initial root.

The new acceptance tests leverage this property to only require the
following:

1. An OCI image that has already been signed and attested by cosign.
2. WireMock recordings to capture the TUF mirror API calls required to
   initialize the TUF root.
3. The root.json representing the trusted initial root.

A script to generate all the required data is also included. This will
be useful if we need to re-generate any of the data as needed. But it is
not required as part of the regular development workflow.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>