Send summary output to both old & new task results
I'm not sure what consumes this, but to be conservative let's write
both result fields for a while. The plan is to transition to
TEST_OUTPUT and eventually remove the deprecated HACBS_TEST_OUTPUT
field.

Note: I didn't change anything in
pkg/schema/examples/hacbs2_valid.json in this commit. I don't think
it will matter, but we might do that later I guess.

JIRA: HACBS-2160