Add "output should match baseline" cucumber step
Having the long chunks of json output inline in the feature file was
starting to feel messy. This tidies it up a little and also allows
reuse.

This could be done for more of the inline json chunks, but in this
commit it's used just for the happy day scenario, because I want to
reuse that particular baseline in the next commit.