Rule dependencies
A rule can specify a custom annotation `depends_on` with a string or a
list of strings containing the code(s) of the rule(s) that a particular
rule depends on. When a rule with dependents is reported as a failure,
warning or it is skipped, dependent rules are filtered out and not
reported.

Ref. https://issues.redhat.com/browse/HACBS-2208