`data` output format
The `ec validate image` command can now output the data used in policy
evaluation with the `--output data=<path>` parameter. The data is
written to the provided path as a YAML file containing an array of data
passed to each evaluation, equivalent of a source group.

A step has been added to the v-e-c to output the data as a result.

Ref. https://issues.redhat.com/browse/HACBS-2132