Destroy global kind cluster at suite clean up
There's a timing issue that happens in certain environments, e.g.
GitHub. The issue seems to appear whenever the number of test cases
using the kind cluster is higher than the number of CPU cores on the
underlying system. (The concurrency level is set to the number of CPU
cores.) By default, GitHub uses a machine that has 2 CPU cores. Thus,
having 4 tests triggers the issue. The first 2 tests complete then
decrease the cluster usage counter which triggers the cluster to be
destroyed. When the next tests start, the cluster is no longer
available.

This commit changes how the global kind cluster is destroyed. It is now
only destroyed at the end of a test suite - once all tests have
finished.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>