Provide certificate information to policy rules
This expands the input file provided to Conftest by adding a `image`
key, in addition to the existing `attestations` key. Which holds the
`ref` and the `signatures key that in turn holds the `keyid`, `sig`,
`certificate`, `chain` and `metadata` keys. For example, to access
the certificate of the signer one could do something like:
`input.image.signatures[_].certificate` in Rego.

The attestation signature information and metadata that was previously
present in the output has been replaced with the image signature
information.

Ref. https://issues.redhat.com/browse/HACBS-2075