Introduce the concept of validators
The idea is to move the various checks performed by the
`ec validate image` command into their own package under validators. By
doing so we enabled a couple of things:

1. We can eventually tie validators to source groups. This would allow
   a policy configuration to determine exactly which validators should
   be executed.

2. We can also add many more validators that rely on specific data
   without having to worry about wasted resources when policies don't
   leverage such data. A validator is reponsible for fetching the data
   it needs. If it's not configured, the data is not fetched.

There's still a lot to go. This commit introduces a single validator
responsible for verifying image signatures. The beginning of a slsa
provenance validator is added for inspiration.

There are some subtle changes here that made sense to do at this time.
For example, some of the logic was moved away from the output package
and into the validator. This was intentional as I believe that is a
better home.

The abstraction is still not perfect. For instance, we should avoid
exposing conftest and consign structs everywhere. I decided to avoid
tackling this existing problem to avoid a larger patch. But it's
something we should take care of at some point.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>