Look at both name and digest annotations
The `org.opencontainers.image.base.name` annotation is meant to hold the
repository of where the parent image is located. While the
`org.opencontainers.image.base.digest` annotation holds the digest of the
parent image.

Previously, the code incorrectly assumed that the `name` label had both
values. This commit still honors the case where a digest is found in
`name`, but it now correctly leverages the `digest` value when it is
available.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>