Refactor the attestation model
Given the general nature of attestations we were to tightly coupled to
the model of build attestations. With this we can evolve to support
different types of attestations. We were also coupled to the output
needed for the reports in our model, which doesn't allow us to evolve
the model without modifying the output.

I think the `attestation.Attestation` model can be simplified even more,
something that I'll look into as well. Most likely the metadata and the
typed statement are not needed and could be kept internal to the
attestation implementation.

Right now we also equate the statement with attestation, further
refinement could introduce a model for the attestation and statement
separately.